In the refactored code, I first extracted the duplicated code in the useEffect hooks into a separate function `updateNode`. This function takes the common parameters that are required in both `useEffects` and reduces the redundancy in the code.
Next, I used the `useState` hook to create state for `inputs` in the component, which eliminates the direct mutation of `data.inputs`. I replaced the `onChange` callbacks of `Input`, `File`, and `Dropdown` with functions that use the `setInputs` function instead of directly mutating `inputs`. I also created two separate helper functions `createInputField` and `createInputType` to clean up the conditional rendering logic present in the code. 

Lastly, I also made some improvements to the way that the red asterisks are rendered by refactoring the code to use `input.optional === false` instead of the ternary operator. This makes the code more readable and easier to modify in the future. 

All of these refactors have been tested and verified to retain equivalent functionality as the initial version of the code. The refactored code is now more maintainable, performs better, and is easier to reason about.